requirements.txt
# requirements.txt  —  Sistema de Control de Acceso YOLO + FaceNet
#
# Instalar con:
#   pip install -r requirements.txt
#
# Si usas GPU NVIDIA, instala primero PyTorch con CUDA desde:
#   https://pytorch.org/get-started/locally/

# ── Deep Learning ──────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── YOLO (Ultralytics YOLOv8) ──────────────────────────────
ultralytics>=8.0.0

# ── FaceNet con PyTorch ────────────────────────────────────
facenet-pytorch>=2.5.3

# ── Visión por computador ──────────────────────────────────
opencv-python>=4.8.0

# ── Utilidades ─────────────────────────────────────────────
numpy>=1.24.0
Pillow>=10.0.0
